Commit python protobuf generated files to git. When installing packages with PIP there's not a good way to execute actions from setup.py. There's cmdclass but it's complex. It's better to have the protoc output under revision control and update them when necessary when the proto files change.
