+++ setup.py
@@ -14,16 +14,8 @@ It supports TensorFlow, PyTorch, MXNet, and XGBoost on Python 3.6+.
 """
 
 # Standard Library
-import logging
-import os
-import shutil
 import sys
-import tempfile
-import urllib
-import urllib.request
 from datetime import date
-from subprocess import check_call
-from zipfile import ZipFile
 
 # Third Party
 import setuptools
@@ -37,50 +29,7 @@ TESTS_PACKAGES = ["pytest", "torchvision", "pandas"]
 INSTALL_REQUIRES = ["protobuf>=3.6.0", "numpy", "packaging", "boto3>=1.10.32"]
 
 
-def _protoc_bundle():
-    import platform
-
-    system = platform.system()
-    machine = platform.machine()
-    if system == "Darwin":
-        archive_url = "https://github.com/protocolbuffers/protobuf/releases/download/v3.11.4/protoc-3.11.4-osx-x86_64.zip"
-    elif system == "Linux":
-        archive_url = f"https://github.com/protocolbuffers/protobuf/releases/download/v3.11.4/protoc-3.11.4-linux-{machine}.zip"
-    else:
-        archive_url = None
-    return archive_url
-
-
-def get_protoc():
-    """make sure protoc is available, otherwise download it and return the path to protoc"""
-    if not shutil.which("protoc"):
-        archive_url = _protoc_bundle()
-        if not archive_url:
-            raise RuntimeError(
-                "protoc not installed and I don't know how to download it, please install manually."
-            )
-        logging.info("Downloading protoc")
-        (fname, headers) = urllib.request.urlretrieve(archive_url)
-        tmpdir = tempfile.mkdtemp(prefix="protoc_smdebug")
-        with ZipFile(fname, "r") as zipf:
-            zipf.extractall(tmpdir)
-        protoc_bin = os.path.join(tmpdir, "bin", "protoc")
-        os.chmod(protoc_bin, 0o755)
-        return protoc_bin
-    return shutil.which("protoc")
-
-
-def compile_protobuf():
-    proto_paths = ["smdebug/core/tfevent/proto"]
-    protoc_bin = get_protoc()
-    for proto_path in proto_paths:
-        proto_files = os.path.join(proto_path, "*.proto")
-        print("compiling protobuf files in {}".format(proto_path))
-    check_call([protoc_bin, proto_path, "--python_out=."])
-
-
 def build_package(version):
-    compile_protobuf()
     packages = setuptools.find_packages(include=["smdebug", "smdebug.*"])
     setuptools.setup(
         name="smdebug",
